% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{area_table}
\alias{area_table}
\title{New stratum table}
\format{
\subsection{\code{area_table}}{
}
}
\usage{
area_table
}
\description{
Consolidates stratum information from all regions (AI, GOA, EBS, NBS so far)
}
\keyword{datasets}
