% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agg_biomass.R
\name{calc_agg_biomass}
\alias{calc_agg_biomass}
\title{Calculate index of total biomass across aggregated strata}
\usage{
calc_agg_biomass(
  biomass_strata = NULL,
  region = c("EBS_STANDARD", "EBS_PLUSNW", "NBS", "GOA", "AI")[1]
)
}
\arguments{
\item{biomass_strata}{a dataframe of stratum biomass, result object from
\code{calc_biomass_stratum()}}

\item{region}{character string. One of c("EBS_STANDARD", "EBS_PLUSNW",
"NBS", "GOA", "AI")}
}
\value{
dataframe of biomass and population abundance estimates across
subareas and across the region, along with variances (CIs to be
added when we can figure out how to do them...)
}
\description{
Calculate index of total biomass across aggregated strata
}
