% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agg_size_comp.R
\name{calc_agg_size_comp}
\alias{calc_agg_size_comp}
\title{Calculate size composition across aggregated subareas}
\usage{
calc_agg_size_comp(
  size_comps,
  region = c("EBS_STANDARD", "EBS_PLUSNW", "NBS", "GOA", "AI")[1]
)
}
\arguments{
\item{size_comps}{a dataframe of stratum biomass, result object from either
\code{AFSC.GAP.DBE::calc_size_stratum_AIGOA()} or
\code{AFSC.GAP.DBE::calc_size_stratum_BS()}}

\item{region}{character string. One of c("EBS_STANDARD", "EBS_PLUSNW",
"NBS", "GOA", "AI")}
}
\value{
dataframe of size composition estimates across
subareas and across region.
}
\description{
Calculate size composition across aggregated subareas
}
