% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_stratum.R
\name{calc_biomass_stratum}
\alias{calc_biomass_stratum}
\title{Calculate index of total biomass per stratum}
\usage{
calc_biomass_stratum(
  gapdata = NULL,
  racebase_tables = lifecycle::deprecated(),
  cpue = NULL
)
}
\arguments{
\item{gapdata}{data object created from \code{gapindex::get_data()}}

\item{racebase_tables}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{gapdata} argument instead.}

\item{cpue}{object created from \code{gapindex::calc_cpue()}.}
}
\value{
\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
N_COUNT              \tab Total number of hauls with positive count data.                                                                                                                                                                                                                                                                                                                                                      \cr
N_HAUL               \tab Total number of hauls.                                                                                                                                                                                                                                                                                                                                                                               \cr
N_LENGTH             \tab Total number of hauls with length data.                                                                                                                                                                                                                                                                                                                                                              \cr
N_WEIGHT             \tab Total number of hauls with positive catch biomass.                                                                                                                                                                                                                                                                                                                                                   \cr
POPULATION_COUNT     \tab The estimated population caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                                                                                                     \cr
POPULATION_VAR       \tab The estimated population variance caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                                                                                            \cr
SPECIES_CODE         \tab The species code of the organism associated with the common_name and scientific_name columns. For a complete species list, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.                                                                                                                                    \cr
STRATUM              \tab RACE database statistical area for analyzing data. Strata were designed using bathymetry and other geographic and habitat-related elements. The strata are unique to each survey region. Stratum of value 0 indicates experimental tows.                                                                                                                                                             \cr
SURVEY               \tab Name and description of survey. The column survey is associated with the srvy and survey_definition_id columns.                                                                                                                                                                                                                                                                                      \cr
SURVEY_DEFINITION_ID \tab The survey definition ID key code is an integer that uniquely identifies a survey region/survey design. The column survey_definition_id is associated with the srvy and survey columns. Full list of survey definition IDs are in RACE_DATA.SURVEY_DEFINITIONS and in the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.\cr
YEAR                 \tab Year the observation (survey) was collected.                                                                                                                                                                                                                                                                                                                                                         \cr
BIOMASS_MT           \tab The estimated total biomass.                                                                                                                                                                                                                                                                                                                                                                         \cr
BIOMASS_VAR          \tab The estimated variance associated with the total biomass.                                                                                                                                                                                                                                                                                                                                            \cr
CPUE_KGKM2_MEAN      \tab The mean catch weight (kilograms) per unit effort (area swept by the net, units squared kilometers).                                                                                                                                                                                                                                                                                                 \cr
CPUE_KGKM2_VAR       \tab The variance of mean catch weight (kilograms) per unit effort (area swept by the net, units squared kilometers).                                                                                                                                                                                                                                                                                     \cr
CPUE_NOKM2_MEAN      \tab The mean of numerical catch per unit effort (area swept by the net, units square kilometers).                                                                                                                                                                                                                                                                                                        \cr
CPUE_NOKM2_VAR       \tab The variance of mean numerical catch per unit effort (area swept by the net, units square kilometers).
}
}
\description{
Calculate index of total biomass per stratum
}
