% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cpue.R
\name{calc_cpue}
\alias{calc_cpue}
\title{Calculate CPUEs at haul level}
\usage{
calc_cpue(racebase_tables = NULL)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}
}
\value{
dataframe of weight and numerical CPUE for the region, species, and
years pulled from \code{gapindex::get_data()}. A table of column name
descriptions is coming soon.
}
\description{
This function zero-fills the catch data and converts numbers
and weights to CPUE.
}
