% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_size_stratum_AIGOA.R
\name{calc_size_stratum_AIGOA}
\alias{calc_size_stratum_AIGOA}
\title{Calculate numbers at length for each stratum}
\usage{
calc_size_stratum_AIGOA(
  racebase_tables = NULL,
  racebase_cpue = NULL,
  racebase_stratum_popn = NULL
)
}
\arguments{
\item{racebase_tables}{data object created from \code{gapindex::get_data()}.}

\item{racebase_cpue}{object created from \code{gapindex::calc_cpue()}.}

\item{racebase_stratum_popn}{a dataframe of stratum abundances, result
object from  \code{gapindex::calc_biomass_stratum()}}
}
\description{
Uses Equations 16 and 17 in Wakabayashi et al. 1985 to calculate
numbers at length. To be used only for the GOA or AI regions.
}
