% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{design_table}
\alias{design_table}
\title{New design table}
\format{
\subsection{\code{design_table}}{
}
}
\usage{
design_table
}
\description{
Pulls which strata to use based on the year and survey. Allows for multiple
stratification designs to be used within a survey, e.g., historical and
2025 GOA strata.
}
\keyword{datasets}
