% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapindex-package.R
\docType{package}
\name{gapindex-package}
\alias{gapindex}
\alias{gapindex-package}
\title{gapindex: Standard AFSC GAP Product Calculations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains functions to calculate standard NOAA AFSC GAP summer groundfish bottom trawl survey data products (CPUE, Biomass, Size Composition, and Age Composition) for the Gulf of Alaska (1990-current), Aleutian Islands (1991-current), Eastern Bering Sea Shelf (1982-current), Northern Bering Sea Shelf (2010-current), and Eastern Bering Sea Slope (2002-2016).
}
\author{
\strong{Maintainer}: Zack Oyafuso \email{zack.oyafuso@noaa.gov}

}
\keyword{internal}
