% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agg_biomass.R
\name{get_agg_biomass}
\alias{get_agg_biomass}
\title{Calculate index of total biomass across aggregated strata}
\usage{
get_agg_biomass(
  biomass_strata = NULL,
  region = c("EBS_STANDARD", "EBS_NW", "NBS", "GOA", "AI")[2]
)
}
\arguments{
\item{biomass_strata}{a dataframe of stratum biomass, result object from
\code{get_biomass_stratum()}}

\item{region}{character string. One of c("EBS_STANDARD", "EBS_NW",
"NBS", "GOA", "AI")}
}
\description{
Calculate index of total biomass across aggregated strata
}
