% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_biomass_stratum.R
\name{get_biomass_stratum}
\alias{get_biomass_stratum}
\title{Calculate index of total biomass per stratum}
\usage{
get_biomass_stratum(
  cpue = NULL,
  haul = NULL,
  strata = NULL,
  vulnerability = 1,
  ci_val = 95
)
}
\arguments{
\item{cpue}{object created from AFSC.GAP.DBE::get_cpue().}

\item{haul}{object created from the haul slot from the object created from
AFSC.GAP.DBE::get_data().}

\item{strata}{object created from the stratum slot from the object created
from AFSC.GAP.DBE::get_data().}

\item{vulnerability}{the vulnerability of the species to the survey
(defaults to 1).}

\item{ci_val}{numeric, ci_val\% confidence interval. Defaults to 95.}
}
\value{
dataframe of biomass and population abundance estimates across
strata, along with variances and CIs.
}
\description{
Calculate index of total biomass per stratum
}
