% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpue.R
\name{get_cpue}
\alias{get_cpue}
\title{Calculate CPUEs at haul level}
\usage{
get_cpue(racebase_tables = NULL)
}
\arguments{
\item{racebase_tables}{object created from AFSC.GAP.DBE::get_data()}
}
\value{
dataframe of weight and numerical CPUE for the region, species, and
years pulled from AFSC.GAP.DBE::get_data()
}
\description{
This function zero-fills the catch data and converts numbers
and weights to CPUE.
}
