% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_info.R
\name{get_station_info}
\alias{get_station_info}
\title{Pull station/grid data}
\usage{
get_station_info(region = NULL, sql_channel = NULL)
}
\arguments{
\item{region}{Survey region (GOA, AI, or EBS)}

\item{sql_channel}{connection created via gapindex::get_connected()}
}
\value{
a dataframe containing station information from the specified survey.
If you request data for the EBS survey, the columns in the dataframe will be STATIONID,
LATITUDE, LONGITUDE, and STRATUM. For AI and GOA, the columns will be AIGRID_ID,
TRAWLABLE, STRATUM, STATIONID, CENTER_LAT, CENTER_LONG, SOUTH_LAT, EAST_LONG, WEST_LONG.
}
\description{
Pull station/grid data
}
