% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_stratum_groupings}
\alias{new_stratum_groupings}
\title{Stratum groupings}
\format{
\subsection{\code{new_stratum_groupings}}{
}
}
\usage{
new_stratum_groupings
}
\description{
DF of which strata are contained in each subarea
}
\keyword{datasets}
