% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{new_stratum_table}
\alias{new_stratum_table}
\title{New stratum table}
\format{
\subsection{\code{new_stratum_table}}{
}
}
\usage{
new_stratum_table
}
\description{
Consolidates stratum information from all regions (AI, GOA, EBS, NBS so far)
}
\keyword{datasets}
