% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_temp_tables.R
\name{remove_temp_tables}
\alias{remove_temp_tables}
\title{Drop Temporary Tables Created within the get_data() call}
\usage{
remove_temp_tables(channel = NULL)
}
\arguments{
\item{channel}{connection to Oracle created via gapindex::get_connected()}
}
\description{
Drop Temporary Tables Created within the get_data() call
}
\keyword{internal}
