% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_query.R
\name{sql_query}
\alias{sql_query}
\title{Streamlined querying function using the DBI package}
\usage{
sql_query(channel, query)
}
\arguments{
\item{channel}{A database connection either made via DBI::dbConnect(), odbc::dbConnect or  RODBC::odbcConnect()}

\item{query}{A character string containing a SQL query.}
}
\value{
if the query is successfully executed, a dataframe of the query will result. Otherwise, an error message is returned as a string (or set of strings).
}
\description{
This function executes a SQL query regardless if the database connection was created via RODBC or DBI/ODBC. combines the DBI::dbSendQuery() and DBI::dbFetch() functions
}
