% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch_entries.R
\name{stitch_entries}
\alias{stitch_entries}
\title{Helper function to format sql queries}
\usage{
stitch_entries(stitch_what = "")
}
\arguments{
\item{stitch_what}{string of values to stitch}
}
\value{
a single string that encapsulates the values of \code{stitch_what} in
a parenthesis.
}
\description{
Helper function to format sql queries
}
