% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stratum_groupings}
\alias{stratum_groupings}
\title{Stratum groupings}
\format{
\subsection{\code{stratum_groupings}}{
}
}
\usage{
stratum_groupings
}
\description{
DF of which strata are contained in each subarea
}
\keyword{datasets}
