% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_oracle.R
\name{upload_oracle}
\alias{upload_oracle}
\title{Upload a table to Oracle with associated metadata}
\usage{
upload_oracle(
  x = NULL,
  table_name = NULL,
  metadata_column = NULL,
  table_metadata = NULL,
  channel = NULL,
  schema = NULL,
  update_table = TRUE,
  update_metadata = TRUE,
  share_with_all_users = TRUE
)
}
\arguments{
\item{x}{Either a character string of the path to the .csv file or a data.frame}

\item{table_name}{Name of the table. (Add error checks to make sure table name follows any peculiarities of Oracle tables)}

\item{metadata_column}{data.frame describing the metadata for each of the
fields in the table. Must contain these columns:
1) colname: name of field
2) colname_long: longer version of name for printing
purposes.
3) units: units of field
4) dataype: Oracle data type
5) colname_desc: Full description of field}

\item{table_metadata}{Description of what the table is}

\item{channel}{Establish your oracle connection using a function
like gapindex::get_connected()}

\item{schema}{character string. The name of the schema to save table.
"GAP_PRODUCTS" is the schema where production tables will live.}

\item{update_table}{boolean. Default = TRUE. Save or drop and save the
table in Oracle.}

\item{update_metadata}{boolean. Default = TRUE. Add table and column
metadata to the tables.}

\item{share_with_all_users}{boolean. Default = TRUE. Give all users in
oracle view permissions.}
}
\value{

}
\description{
T
}
