% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agecomp_region.R
\name{calc_agecomp_region}
\alias{calc_agecomp_region}
\title{Calculate region-level age composition and mean/std.dev length at age}
\usage{
calc_agecomp_region(racebase_tables, age_comps_stratum)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{age_comps_stratum}{a named list of stratum age comps and numbers by
survey/year/stratum/sex/length, a result object
from \code{gapindex::calc_agecomp_stratum()}.}
}
\value{
dataframe of age composition and mean/standard deviation of length
at age aggregated across regions.
}
\description{
Calculate region-level age composition and mean/std.dev length at age
}
