% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agecomp_region.R
\name{calc_agecomp_region}
\alias{calc_agecomp_region}
\title{Calculate region-level age composition and mean/std.dev length at age}
\usage{
calc_agecomp_region(racebase_tables, age_comps_stratum)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{age_comps_stratum}{a named list of stratum age comps and numbers by
survey/year/stratum/sex/length, a result object
from \code{gapindex::calc_agecomp_stratum()}.}
}
\value{
dataframe of age composition and mean/standard deviation of length
at age aggregated across regions.

\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
YEAR                 \tab Year the survey was conducted in.                                                                                                                                                                                                                                                                                          \cr
AGE                  \tab Age bin of a taxon in years estimated by the age comp estimate.                                                                                                                                                                                                                                                            \cr
AREA_ID              \tab Area ID code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.                                                                                                                                               \cr
LENGTH_MM_MEAN       \tab Mean length estimated in age comp estimate.                                                                                                                                                                                                                                                                                \cr
LENGTH_MM_SD         \tab Variance of mean length estimated in age comp estimate.                                                                                                                                                                                                                                                                    \cr
POPULATION_COUNT     \tab The estimated population caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                           \cr
SEX                  \tab Sex of a specimen where "1" = "Male", "2" = "Female", "3" = Unsexed.                                                                                                                                                                                                                                                       \cr
SPECIES_CODE         \tab The species code of the organism associated with the 'common_name' and 'scientific_name' columns. For a complete species list, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.                                                      \cr
SURVEY               \tab Name and description of survey. The column 'survey' is associated with the 'srvy' and 'survey_id' columns.                                                                                                                                                                                                                 \cr
SURVEY_DEFINITION_ID \tab This number uniquely identifies a survey. Name and description of survey. The column 'survey_id' is associated with the 'srvy' and 'survey' columns. For a complete list of surveys, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.
}
}
\description{
Calculate region-level age composition and mean/std.dev length at age
}
