% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agecomp_region.R
\name{calc_agecomp_region}
\alias{calc_agecomp_region}
\title{Calculate region-level age composition and mean/std.dev length at age}
\usage{
calc_agecomp_region(
  gapdata = NULL,
  racebase_tables = lifecycle::deprecated(),
  agecomp_stratum = NULL,
  age_comps_stratum = lifecycle::deprecated()
)
}
\arguments{
\item{gapdata}{data object created from \code{gapindex::get_data()}}

\item{racebase_tables}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{gapdata} argument instead.}

\item{agecomp_stratum}{a named list of stratum age comps and numbers by
survey/year/stratum/sex/length, a result object
from \code{gapindex::calc_agecomp_stratum()}.}

\item{age_comps_stratum}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{agecomp_stratum} argument instead.}
}
\value{
dataframe of age composition and mean/standard deviation of length
at age aggregated across regions.

\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
LENGTH_MM_MEAN       \tab Mean length (millimeters)                                                                                                                                                                                                                                                                                                                                                                            \cr
LENGTH_MM_SD         \tab Variance of mean length.                                                                                                                                                                                                                                                                                                                                                                             \cr
POPULATION_COUNT     \tab The estimated population caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                                                                                                     \cr
SEX                  \tab Sex of a specimen where "1" = "Male", "2" = "Female", "3" = Unsexed.                                                                                                                                                                                                                                                                                                                                 \cr
SPECIES_CODE         \tab The species code of the organism associated with the common_name and scientific_name columns. For a complete species list, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.                                                                                                                                    \cr
SURVEY_DEFINITION_ID \tab The survey definition ID key code is an integer that uniquely identifies a survey region/survey design. The column survey_definition_id is associated with the srvy and survey columns. Full list of survey definition IDs are in RACE_DATA.SURVEY_DEFINITIONS and in the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.\cr
YEAR                 \tab Year the observation (survey) was collected.                                                                                                                                                                                                                                                                                                                                                         \cr
AGE                  \tab Age bin of taxon. Age bin of a taxon in years estimated by the age comp estimate. Age -9 indicates unaged lengths for a particular sex because no otoliths were collected for that sex/length combination. Age -99 indicates a case where no lengths were collected within a stratum for a species/year even though catch numbers were recorded.                                                     \cr
AREA_ID_FOOTPRINT    \tab Survey footprint, usually equivalent to the SURVEY_DEFINITION_ID with the exception of the Standard and Standard +NW survey footprints in the Eastern Bering Sea shelf bottom trawl survey                                                                                                                                                                                                           \cr
AREA_ID              \tab Area ID key code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.
}
}
\description{
Calculate region-level age composition and mean/std.dev length at age
}
