% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agecomp_stratum.R
\name{calc_agecomp_stratum}
\alias{calc_agecomp_stratum}
\title{Calculate stratum-level age composition and mean/std.dev length at age}
\usage{
calc_agecomp_stratum(racebase_tables = NULL, size_comp = NULL)
}
\arguments{
\item{racebase_tables}{data object created from \code{gapindex::get_data()}}

\item{size_comp}{a dataframe of stratum abundances, result
object from either \code{calc_sizecomp_aigoa_stratum()}
or \code{calc_sizecomp_bs_stratum()} depending on the region.}
}
\description{
Calculate stratum-level age composition and mean/std.dev length at age
}
