% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agecomp_stratum.R
\name{calc_agecomp_stratum}
\alias{calc_agecomp_stratum}
\title{Calculate stratum-level age composition and mean/std.dev length at age}
\usage{
calc_agecomp_stratum(racebase_tables = NULL, size_comp = NULL)
}
\arguments{
\item{racebase_tables}{data object created from \code{gapindex::get_data()}}

\item{size_comp}{a dataframe of stratum abundances, result
object from either \code{calc_sizecomp_aigoa_stratum()}
or \code{calc_sizecomp_bs_stratum()} depending on the region.}
}
\value{
A named list with two elements. "age_comp" is a dataframe of numbers
at age by survey, year, stratum, species, and sex. "length_at_age" is a
support table for the gapindex::calc_agecomp_region() function and should
not be used.\tabular{ll}{
   Field Name \tab Description \cr
   SURVEY_DEFINITION_ID \tab Integer number identifier corresponding to survey region. See   gapindex::survey_ids for a list of relevant survey regions. \cr
   SURVEY \tab Survey region. \cr
   STRATUM \tab Stratum ID. STRATUM = 0 indicates an experimental tow. \cr
   YEAR \tab Year the survey was conducted in. \cr
   SPECIES_CODE \tab Taxon code. \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{See the code book for the full list.}. \cr
   SEX \tab Sex of a specimen where "1" = "Male", "2" =   "Female", "3" = Unsexed. \cr
   AGE \tab Age (years). \cr
   POPULATION_COUNT \tab Total number of individuals. \cr
   LENGTH_MM_MEAN \tab Estimated mean length-at-age (mm) weighted by numbers-at-length. \cr
   LENGTH_MM_SD \tab Standard deviation associated with the estimated mean length-at-age. \cr
}
}
\description{
Calculate stratum-level age composition and mean/std.dev length at age
}
