% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agg_biomass.R
\name{calc_agg_biomass}
\alias{calc_agg_biomass}
\title{Calculate index of total biomass across aggregated subareas}
\usage{
calc_agg_biomass(racebase_tables, biomass_strata = NULL)
}
\arguments{
\item{racebase_tables}{data object created from `AFSC.GAP.DBE::get_data()``}

\item{biomass_strata}{a dataframe of stratum biomass, result object from
\code{AFSC.GAP.DBE::calc_biomass_stratum()}}
}
\value{
dataframe of biomass and population abundance estimates across
subareas and across the region, along with variances.
}
\description{
Calculate index of total biomass across aggregated subareas
}
