% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_agg_size_comp.R
\name{calc_agg_size_comp}
\alias{calc_agg_size_comp}
\title{Calculate size composition across aggregated subareas}
\usage{
calc_agg_size_comp(racebase_tables, size_comps)
}
\arguments{
\item{racebase_tables}{data object created from `AFSC.GAP.DBE::get_data()``}

\item{size_comps}{a dataframe of stratum biomass, result object from either
\code{AFSC.GAP.DBE::calc_size_stratum_AIGOA()} or
\code{AFSC.GAP.DBE::calc_size_stratum_BS()}}
}
\value{
dataframe of size composition estimates across
subareas and across region.
}
\description{
Calculate size composition across aggregated subareas
}
