% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_stratum.R
\name{calc_biomass_stratum}
\alias{calc_biomass_stratum}
\title{Calculate index of total biomass per stratum}
\usage{
calc_biomass_stratum(racebase_tables = NULL, cpue = NULL, vulnerability = 1)
}
\arguments{
\item{racebase_tables}{data object created from \code{gapindex::get_data()}}

\item{cpue}{object created from \code{gapindex::calc_cpue()}.}

\item{vulnerability}{the vulnerability of the species to the survey
(defaults to 1).}
}
\value{
dataframe of biomass and population abundance estimates across
strata, along with variances and CIs.
}
\description{
Calculate index of total biomass per stratum
}
