% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_subarea.R
\name{calc_biomass_subarea}
\alias{calc_biomass_subarea}
\title{Calculate index of total biomass across aggregated subareas}
\usage{
calc_biomass_subarea(
  racebase_tables = deprecated(),
  gapdata = NULL,
  biomass_strata = deprecated(),
  biomass_stratum = NULL
)
}
\arguments{
\item{racebase_tables}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{gapdata} argument instead.}

\item{gapdata}{data object created from \code{gapindex::get_data()}}

\item{biomass_strata}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{biomass_stratum} argument instead.}

\item{biomass_stratum}{a dataframe of stratum biomass, result object from
\code{gapindex::calc_biomass_stratum()}}
}
\value{
\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
YEAR                 \tab Year the survey was conducted in.                                                                                                                                                                                                                                                                                          \cr
AREA_ID              \tab Area ID code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.                                                                                                                                               \cr
BIOMASS_MT           \tab The estimated biomass caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                              \cr
BIOMASS_VAR          \tab The estimated biomass variance caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                     \cr
N_WEIGHT             \tab Total number of hauls with positive catch/weighed taxon data used in calculation.                                                                                                                                                                                                                                          \cr
N_COUNT              \tab Total number of hauls with positive taxon counts used in calculation.                                                                                                                                                                                                                                                      \cr
N_HAUL               \tab Total number of valid hauls used in calculation.                                                                                                                                                                                                                                                                           \cr
N_LENGTH             \tab Total number of hauls with taxon length data used in calculation.                                                                                                                                                                                                                                                          \cr
CPUE_KGKM2_MEAN      \tab The mean of catch weight (kilograms) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.                                                                                                                                                                                       \cr
CPUE_KGKM2_VAR       \tab The variance of mean of catch weight (kilograms) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.                                                                                                                                                                           \cr
CPUE_NOKM2_MEAN      \tab The mean of catch count (number) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.                                                                                                                                                                                           \cr
CPUE_NOKM2_VAR       \tab The variance of mMean of catch count (number) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.                                                                                                                                                                              \cr
POPULATION_COUNT     \tab The estimated population caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                           \cr
POPULATION_VAR       \tab The estimated population variance caught in the survey for a species, group, or total for a given survey.                                                                                                                                                                                                                  \cr
SPECIES_CODE         \tab The species code of the organism associated with the 'common_name' and 'scientific_name' columns. For a complete species list, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.                                                      \cr
SURVEY               \tab Name and description of survey. The column 'survey' is associated with the 'srvy' and 'survey_id' columns.                                                                                                                                                                                                                 \cr
SURVEY_DEFINITION_ID \tab This number uniquely identifies a survey. Name and description of survey. The column 'survey_id' is associated with the 'srvy' and 'survey' columns. For a complete list of surveys, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.
}
}
\description{
Calculate index of total biomass across aggregated subareas
}
