% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_subarea.R
\name{calc_biomass_subarea}
\alias{calc_biomass_subarea}
\title{Calculate index of total biomass across aggregated subareas}
\usage{
calc_biomass_subarea(racebase_tables = NULL, biomass_strata = NULL)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{biomass_strata}{a dataframe of stratum biomass, result object from
\code{gapindex::calc_biomass_stratum()}}
}
\value{
dataframe of biomass and population abundance estimates across
subareas and across the region, along with variances.
}
\description{
Calculate index of total biomass across aggregated subareas
}
