% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_biomass_subarea.R
\name{calc_biomass_subarea}
\alias{calc_biomass_subarea}
\title{Calculate index of total biomass across aggregated subareas}
\usage{
calc_biomass_subarea(racebase_tables = NULL, biomass_strata = NULL)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{biomass_strata}{a dataframe of stratum biomass, result object from
\code{gapindex::calc_biomass_stratum()}}
}
\value{
dataframe of biomass and population abundance estimates (with
associated variances) across survey, year, species, and subarea (AREA_ID).\tabular{ll}{
   Field Name \tab Description \cr
   SURVEY_DEFINITION_ID \tab Integer number identifier corresponding to survey region. See   gapindex::survey_ids for a list of relevant survey regions. \cr
   SURVEY \tab Survey region. \cr
   AREA_ID \tab Integer identifier for a subarea. See   gapindex::area_table for the full list. \cr
   SPECIES_CODE \tab Taxon code. \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{See the code book for the full   list.}. \cr
   YEAR \tab Survey year. \cr
   N_HAUL \tab Number of hauls in stratum. \cr
   N_WEIGHT \tab Number of hauls with a positive catch weight. \cr
   N_COUNT \tab Number of hauls with a positive numerical catch. \cr
   N_LENGTH \tab Number of hauls with length-frequency data \cr
   CPUE_KGKM2_MEAN \tab Mean weight catch per area swept (kg/km^2). \cr
   CPUE_KGKM2_VAR \tab Variance of the mean weight catch per area swept. \cr
   CPUE_NOKM2_MEAN \tab Mean numerical catch per area swept (no/km^2). \cr
   CPUE_NOKM2_VAR \tab Variance of the mean weight catch per area swept \cr
   BIOMASS_MT \tab Estimated total biomass (mt). \cr
   BIOMASS_VAR \tab Variance associated with the estimated total biomass. \cr
   POPULATION_COUNT \tab Estimated total numerical abundance (numbers). \cr
   POPULATION_VAR \tab Variance associated with the estimated total numerical abundance. \cr
}

dataframe of biomass and population abundance estimates across
subareas and across the region, along with variances.
}
\description{
Calculate index of total biomass across aggregated subareas
}
