% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cpue.R
\name{calc_cpue}
\alias{calc_cpue}
\title{Calculate haul-level catch per unit effort}
\usage{
calc_cpue(gapdata = NULL, racebase_tables = lifecycle::deprecated())
}
\arguments{
\item{gapdata}{data object created from \code{gapindex::get_data()}}

\item{racebase_tables}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Use the
\code{gapdata} argument instead.}
}
\value{
\tabular{ll}{
\bold{Field} \tab \bold{Description} \cr
DEPTH_M              \tab Bottom depth (meters).                                                                                                                                                                                                                                                                                                                                                                               \cr
DESIGN_YEAR          \tab Year ID associated with a given value AREA_ID. This field describes the changes in the survey design over time.                                                                                                                                                                                                                                                                                      \cr
HAULJOIN             \tab This is a unique numeric identifier assigned to each (vessel, cruise, and haul) combination.                                                                                                                                                                                                                                                                                                         \cr
LATITUDE_DD_END      \tab Latitude (one hundred thousandth of a decimal degree) of the end of the haul.                                                                                                                                                                                                                                                                                                                        \cr
LATITUDE_DD_START    \tab Latitude (one hundred thousandth of a decimal degree) of the start of the haul.                                                                                                                                                                                                                                                                                                                      \cr
LONGITUDE_DD_END     \tab Longitude (one hundred thousandth of a decimal degree) of the end of the haul.                                                                                                                                                                                                                                                                                                                       \cr
LONGITUDE_DD_START   \tab Longitude (one hundred thousandth of a decimal degree) of the start of the haul.                                                                                                                                                                                                                                                                                                                     \cr
SPECIES_CODE         \tab The species code of the organism associated with the common_name and scientific_name columns. For a complete species list, review the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.                                                                                                                                    \cr
STRATUM              \tab RACE database statistical area for analyzing data. Strata were designed using bathymetry and other geographic and habitat-related elements. The strata are unique to each survey region. Stratum of value 0 indicates experimental tows.                                                                                                                                                             \cr
SURVEY               \tab Name and description of survey. The column survey is associated with the srvy and survey_definition_id columns.                                                                                                                                                                                                                                                                                      \cr
SURVEY_DEFINITION_ID \tab The survey definition ID key code is an integer that uniquely identifies a survey region/survey design. The column survey_definition_id is associated with the srvy and survey columns. Full list of survey definition IDs are in RACE_DATA.SURVEY_DEFINITIONS and in the \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{code books}.\cr
WEIGHT_KG            \tab Weight (thousandths of a kilogram) of individuals in a haul by taxon.                                                                                                                                                                                                                                                                                                                                \cr
YEAR                 \tab Year the observation (survey) was collected.                                                                                                                                                                                                                                                                                                                                                         \cr
AREA_SWEPT_KM2       \tab The area the net covered while the net was fishing (kilometers squared), defined as the distance fished times the net width.                                                                                                                                                                                                                                                                         \cr
BOTTOM_TEMPERATURE_C \tab Bottom temperature (tenths of a degree Celsius); NA indicates removed or missing values.                                                                                                                                                                                                                                                                                                             \cr
COUNT                \tab Total whole number of individuals caught in haul or samples collected.                                                                                                                                                                                                                                                                                                                               \cr
CPUE_KGKM2           \tab Catch weight (kilograms) per unit effort (area swept by the net, units square kilometers).                                                                                                                                                                                                                                                                                                           \cr
CPUE_NOKM2           \tab Numerical catch per unit effort (area swept by the net, units square kilometers).                                                                                                                                                                                                                                                                                                                    \cr
CRUISE               \tab This is a six-digit integer identifying the cruise number of the form: YYYY99 (where YYYY = year of the cruise; 99 = 2-digit number and is sequential; 01 denotes the first cruise that vessel made in this year, 02 is the second, etc.).                                                                                                                                                           \cr
CRUISEJOIN           \tab Unique integer ID assigned to each survey, vessel, and year combination.
}
}
\description{
This function calculates and zero-fills weight and numerical
catch per unit effort (area swept, km2)
}
