% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sizecomp_stratum.R
\name{calc_sizecomp_stratum}
\alias{calc_sizecomp_stratum}
\title{Calculate numerical CPUE at length by haul}
\usage{
calc_sizecomp_stratum(
  racebase_tables = NULL,
  racebase_cpue = NULL,
  racebase_stratum_popn = NULL,
  spatial_level = c("stratum", "haul")[1],
  fill_NA_method = c("AIGOA", "BS")[1]
)
}
\arguments{
\item{racebase_tables}{data object created from \code{gapindex::get_data()}.}

\item{racebase_cpue}{object created from \code{gapindex::calc_cpue()}.}
}
\value{
dataframe of numerical CPUE ("S_ijklm_NO_KM2") by survey ("SURVEY"),
year ("YEAR"), stratum ("STRATUM"), haul ("HAULJOIN), species (SPECIES_CODE),
sex (SEX), and length (LENGTH_MM).
}
\description{
Uses Equation 16 in Wakabayashi et al. 1985 to calculate
numbers-catch-per-unit-effort (CPUE, Numbers/km2) at length.
}
