% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sizecomp_subarea.R
\name{calc_sizecomp_subareas}
\alias{calc_sizecomp_subareas}
\title{Calculate size composition across aggregated subareas}
\usage{
calc_sizecomp_subareas(racebase_tables, size_comps)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{size_comps}{a dataframe of stratum biomass, result object from either
\code{gapindex::calc_sizecomp_aigoa_stratum()} or
\code{gapindex::calc_sizecomp_bs_stratum()}}
}
\value{
dataframe of size composition estimates across
subareas and across region.
}
\description{
Calculate size composition across aggregated subareas
}
