% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sizecomp_subarea.R
\name{calc_sizecomp_subareas}
\alias{calc_sizecomp_subareas}
\title{Calculate size composition across aggregated subareas}
\usage{
calc_sizecomp_subareas(racebase_tables, size_comps)
}
\arguments{
\item{racebase_tables}{data object created from `gapindex::get_data()``}

\item{size_comps}{a dataframe of stratum biomass, result object from either
\code{gapindex::calc_sizecomp_aigoa_stratum()} or
\code{gapindex::calc_sizecomp_bs_stratum()}}
}
\value{
dataframe of numbers-at-length by survey, year, subarea (AREA_ID), species, and sex\tabular{ll}{
   Field Name \tab Description \cr
   SURVEY_DEFINITION_ID \tab Integer number identifier corresponding to survey region. See   gapindex::survey_ids for a list of relevant survey regions. \cr
   SURVEY \tab Survey region. \cr
   YEAR \tab Survey year. \cr
   AREA_ID \tab Integer identifier for a subarea. See   gapindex::area_table for the full list. \cr
   SPECIES_CODE \tab Taxon code. \href{https://www.fisheries.noaa.gov/resource/document/groundfish-survey-species-code-manual-and-data-codes-manual}{See the code book for the full list.}. \cr
   LENGTH_MM \tab Length bin (mm). \cr
   SEX \tab Sex code where "1" = "Male", "2" =   "Female", "3" = Unsexed. \cr
   POPULATION_COUNT \tab Total number of individuals. \cr
}
}
\description{
Calculate size composition across aggregated subareas
}
