% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_one_bootstrap.R
\name{do_one_bootstrap}
\alias{do_one_bootstrap}
\title{Bootstrap a CPUE table from gapindex}
\usage{
do_one_bootstrap(boot, gapdata, gapcpue)
}
\arguments{
\item{boot}{(numeric) bootstrap number.}

\item{gapcpue}{object created from gapindex::calc_cpue(). This can have multiple years of data, but should be one species and one region only.}
}
\value{
a resampled CPUE table with one row per sample.
}
\description{
Bootstrap a CPUE table from gapindex
}
\details{
This function returns one single resample of a CPUE table. Usually, you would be running this function several times (1 for each bootstrap you want to do of the data).
}
\examples{
testdata <- gapindex::get_data(
  year_set = c(1996, 1999),
  survey_set = c("GOA", "AI", "EBS", "NBS", "BSS")[1],
  spp_codes = c(21740, 30060, 10110)[1],
  haul_type = 3,
  abundance_haul = c("Y", "N")[1],
  pull_lengths = FALSE
)
testcpue <- gapindex::calc_cpue(gapdata = testdata)

x1 <- do_one_bootstrap(boot = 1, gapdata = testdata, gapcpue = testcpue)
x2 <- do_one_bootstrap(boot = 2, gapdata = testdata, gapcpue = testcpue)

library(ggplot2)
ggplot(testcpue, aes(x = factor(YEAR), y = CPUE_KGKM2)) +
  geom_jitter(width = 0.05, height = 0.00001) +
  geom_jitter(data = x1, width = 0.05, color = "blue", alpha = 0.2) +
  geom_jitter(data = x2, width = 0.05, color = "red", alpha = 0.2)
}
