% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapindex-package.R
\docType{package}
\name{gapindex-package}
\alias{gapindex}
\alias{gapindex-package}
\title{gapindex: Standard AFSC GAP Product Calculations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package contains functions to calculate standard data products (CPUE, Biomass, Size Composition, and Age Composition) resulting from the NOAA AFSC GAP summer groundfish bottom trawl survey.
}
\author{
\strong{Maintainer}: Zack Oyafuso \email{zack.oyafuso@noaa.gov}

}
\keyword{internal}
