% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_connected.R
\name{get_connected}
\alias{get_connected}
\title{Define RODBC connection to Oracle}
\usage{
get_connected(schema = "AFSC")
}
\arguments{
\item{schema}{A registered data source name. To be used as the \code{dsn} argument in \code{RODBC::odbcConnect()}}
}
\value{
channel of class "RODBC". See \code{?RODBC::odbcConnect()} for more detail
}
\description{
Creates the RODBC connection to Oracle needed to pull SQL
queries from RACE database. Make sure you are connected to the VPN before
running the function.
}
