% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Pull data from RACEBASE}
\usage{
get_data(
  year_set = c(1996, 1999),
  survey_set = c("GOA", "BS", "AI")[1],
  spp_codes = c(`Pacific cod` = 21720, `walleye pollock` = 21740),
  haul_type = 3,
  abundance_haul = c("Y"),
  sql_channel = NULL
)
}
\arguments{
\item{year_set}{numeric or integer vector of years}

\item{survey_set}{character string. One of c("GOA", "BS", "AI")}

\item{spp_codes}{character vector of species codes. See ... for full list}

\item{haul_type}{integer. Defaults to haul type "3" for Standard bottom
sample (preprogrammed station) used for biomass estimation}

\item{abundance_haul}{character string. "Y" are abundance hauls (what does
this mean?) and "N" are other hauls.}

\item{sql_channel}{connection created via AFSC.GAP.DBE::get_connected()}
}
\value{
a named list containing cruise, haul, catch, and stratum information
for the region, years, and species of interest.
}
\description{
Pulls cruise, haul, catch, and stratum information for the
region, years, and species of interest.
}
