% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bootstraps.R
\name{run_bootstraps}
\alias{run_bootstraps}
\title{Bootstrap haul data and return index information}
\usage{
run_bootstraps(nboot = 5, gapdata, gapcpue)
}
\arguments{
\item{nboot}{number of bootstraps}

\item{gapdata}{an object generated by gapindex::get_data(). It should contain gap data for one species and one region. Multiple years ok.}

\item{gapcpue}{an object generated by gapinex::calc_cpue().}
}
\description{
Bootstrap haul data and return index information
}
\examples{
 testdata <- gapindex::get_data(
  year_set = c(1996, 1999),
  survey_set = c("GOA", "AI", "EBS", "NBS", "BSS")[1],
  spp_codes = c(21740, 30060, 10110)[1],
  haul_type = 3,
  abundance_haul = c("Y", "N")[1],
  pull_lengths = TRUE
)
testcpue <- gapindex::calc_cpue(gapdata = testdata)
run_bootstraps(nboot = 5, gapdata = testdata, gapcpue = testcpue)
}
